Interface improvements
