+++ src/main/java/org/roda/rodain/rules/ui/RuleModalStage.java
@@ -28,7 +28,7 @@ public class RuleModalStage extends Stage{
 
         setResizable(true);
 
-        Scene scene = new Scene(new HBox(), 800, 515);
+        Scene scene = new Scene(new HBox(), 800, 535);
         scene.getStylesheets().add(ClassLoader.getSystemResource("css/modal.css").toExternalForm());
         scene.getStylesheets().add(ClassLoader.getSystemResource("css/shared.css").toExternalForm());
         setScene(scene);
